+++ cspslice.h
@@ -14,7 +14,7 @@ public:
 	Complex* p_data = NULL;
 	rectangle region;
 	long memsize;
-	double wavelength;
+	int wavelength;			// nm
 	long2 getDimensions();
 	long getNumberOfElements();
 	virtual int clear() { return 0; };
@@ -29,8 +29,8 @@ class hspslice : public spslice, public hmemory {
 public:
 	hcube* datacube;
 	hspslice() {};
-	hspslice(hcube*, std::valarray<double>, rectangle, double);
-	hspslice(hcube*, std::valarray<Complex>, rectangle, double);
+	hspslice(hcube*, std::valarray<double>, rectangle, int);
+	hspslice(hcube*, std::valarray<Complex>, rectangle, int);
 	hspslice(hcube*, dspslice*);
 	~hspslice();
 	int clear();
@@ -45,8 +45,8 @@ class dspslice : public spslice, public dmemory {
 public:
 	dcube* datacube;
 	dspslice() {};
-	dspslice(dcube*, std::valarray<double>, rectangle, double);
-	dspslice(dcube*, std::valarray<Complex>, rectangle, double);
+	dspslice(dcube*, std::valarray<double>, rectangle, int);
+	dspslice(dcube*, std::valarray<Complex>, rectangle, int);
 	dspslice(dcube*, hspslice*);
 	~dspslice();
 	int clear();
